separate Storage into multiple small interfaces such as EdgeFetcher/VertexMutator, ...
